cloudfunctions/detect_handwriting/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-vision
google-cloud-storage